:sparkles: Add printing the output path and garbage collecting previous outputs
