+++ cmd/prompto/main.go
@@ -1,124 +1,15 @@
 package main
 
 import (
-	"context"
 	"fmt"
-	"github.com/go-go-golems/glazed/pkg/cmds"
-	"github.com/go-go-golems/glazed/pkg/cmds/layers"
-	"github.com/go-go-golems/glazed/pkg/cmds/parameters"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"strings"
-
+	files2 "github.com/go-go-golems/glazed/pkg/helpers/files"
+	"github.com/go-go-golems/prompto/pkg"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
+	"os"
+	"path/filepath"
+	"strings"
 )
-
-type FileType int
-
-const (
-	Plain FileType = iota
-	Executable
-	TemplateCommand
-)
-
-type FileInfo struct {
-	Name    string
-	Type    FileType
-	Command *cmds.TemplateCommand
-}
-
-func loadTemplateCommand(path string) (*cmds.TemplateCommand, bool) {
-	f, err := os.Open(path)
-	if err != nil {
-		return nil, false
-	}
-	defer func(f *os.File) {
-		_ = f.Close()
-	}(f)
-
-	tcl := &cmds.TemplateCommandLoader{}
-	commands, err := tcl.LoadCommandFromYAML(f)
-	if err != nil {
-		return nil, false
-	}
-	if len(commands) != 1 {
-		return nil, false
-	}
-	return commands[0].(*cmds.TemplateCommand), err == nil
-}
-
-func getFilesFromRepo(repo string) ([]FileInfo, error) {
-	promptoDir := filepath.Join(repo, "prompto")
-
-	// check if repo/prompto exists, else return empty list
-	if _, err := os.Stat(promptoDir); os.IsNotExist(err) {
-		return []FileInfo{}, nil
-	}
-	var files []FileInfo
-
-	err := filepath.Walk(promptoDir, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-
-		// skip . files and directories
-		if strings.HasPrefix(info.Name(), ".") {
-			if info.IsDir() {
-				return filepath.SkipDir
-			}
-			return nil
-		}
-
-		// strip the repo path from the file name
-		relpath := path[len(repo):]
-		name := strings.TrimPrefix(relpath, "/prompto/")
-
-		if !info.IsDir() {
-			file := FileInfo{Name: name, Type: Plain}
-			ext := strings.ToLower(filepath.Ext(name))
-
-			isYAMLfile := ext == ".yaml" || ext == ".yml"
-
-			// check for symbolic link
-			if info.Mode()&os.ModeSymlink != 0 {
-				// check if the link source is executable
-				info_, err := os.Stat(path)
-				if err != nil {
-					return err
-				}
-				if (info_.Mode() & 0111) != 0 {
-					file.Type = Executable
-				}
-			} else {
-				if (info.Mode() & 0111) != 0 {
-					file.Type = Executable
-				}
-			}
-
-			if file.Type != Executable {
-				if isYAMLfile {
-					if cmd, ok := loadTemplateCommand(path); ok {
-						file.Name = name[:len(name)-len(ext)]
-						file.Type = TemplateCommand
-						file.Command = cmd
-					}
-				}
-			}
-
-			files = append(files, file)
-		}
-
-		return nil
-	})
-
-	if err != nil {
-		return nil, err
-	}
-
-	return files, nil
-}
 
 func main() {
 	var rootCmd = &cobra.Command{
@@ -127,6 +18,8 @@ func main() {
 		Long: `This program loads a list of repositories from a yaml config file
 and looks for a file that matches the prompt.`,
 	}
+
+	getCmd.Flags().Bool("print-path", false, "Print the path of the prompt")
 
 	rootCmd.AddCommand(getCmd)
 	rootCmd.AddCommand(listCmd)
@@ -160,68 +53,63 @@ func get(cmd *cobra.Command, args []string) error {
 	prompt := args[0]
 	restArgs := args[1:]
 
+	printPath, _ := cmd.Flags().GetBool("print-path")
+
 	repositories := viper.GetStringSlice("repositories")
 
 	for _, repo := range repositories {
-		files, err := getFilesFromRepo(repo)
+		files, err := pkg.GetFilesFromRepo(repo)
 		if err != nil {
 			return err
 		}
 
 		for _, file := range files {
 			if file.Name == prompt {
-				path := filepath.Join(repo, "prompto", file.Name)
-				switch file.Type {
-				case Executable:
-					// The file is executable, execute it
-					c := exec.Command(path, restArgs...)
-					c.Dir = repo
-					c.Stdout = os.Stdout
-					c.Stderr = os.Stderr
 
-					// Get the current working directory
-					currentDir, err := os.Getwd()
+				if printPath {
+					if file.Type == pkg.Plain {
+						fmt.Println(filepath.Join(repo, "prompto", file.Name))
+						continue
+					}
+
+					s, err := pkg.RenderFile(repo, file, restArgs)
 					if err != nil {
-						// Handle error
 						return err
 					}
 
-					// Set the PROMPTO_PARENT_PWD environment variable to the current directory
-					c.Env = append(os.Environ(), "PROMPTO_PARENT_PWD="+currentDir)
+					// transform prompt into a safe filename by removing / and \
+					promptString := strings.ReplaceAll(prompt, "/", "-")
+					promptString = strings.ReplaceAll(promptString, "\\", "-")
 
-					return c.Run()
-				case TemplateCommand:
-					// The file is a glazed TemplateCommand, execute it by passing the arguments
-					buf := &strings.Builder{}
-					parsedLayers := map[string]*layers.ParsedParameterLayer{}
-					ps, args_, err := parameters.GatherFlagsFromStringList(
-						restArgs, file.Command.Flags,
-						false, false,
-						"")
+					deletedFiles, err := files2.GarbageCollectTemporaryFiles(os.TempDir(), "prompto-*", 20)
 					if err != nil {
 						return err
 					}
-					arguments, err := parameters.GatherArguments(args_, file.Command.Arguments, false, false)
+
+					//_, _ = fmt.Fprintf(os.Stderr, "Deleted %d files\n", len(deletedFiles))
+					_ = deletedFiles
+
+					f, err := os.CreateTemp(os.TempDir(), fmt.Sprintf("prompto-%s-*", promptString))
 					if err != nil {
 						return err
 					}
-					for p := arguments.Oldest(); p != nil; p = p.Next() {
-						k, v := p.Key, p.Value
-						ps[k] = v
-					}
-					err = file.Command.RunIntoWriter(context.Background(), parsedLayers, ps, buf)
+
+					_, err = f.Write([]byte(s))
 					if err != nil {
+						_ = f.Close()
 						return err
 					}
-					fmt.Println(buf.String())
-				case Plain:
-					// The file is not executable, print its content
-					b, err := os.ReadFile(path)
-					if err != nil {
-						return err
-					}
-					fmt.Println(string(b))
+					_ = f.Close()
+
+					fmt.Println(f.Name())
+
+					continue
 				}
+				s, err := pkg.RenderFile(repo, file, restArgs)
+				if err != nil {
+					return err
+				}
+				fmt.Println(s)
 			}
 		}
 	}
@@ -233,7 +121,7 @@ func list(cmd *cobra.Command, args []string) error {
 	repositories := viper.GetStringSlice("repositories")
 
 	for _, repo := range repositories {
-		files, err := getFilesFromRepo(repo)
+		files, err := pkg.GetFilesFromRepo(repo)
 		if err != nil {
 			return err
 		}
